move fn from utils.ts -> desktop_notifications.ts
