+++ webapp/src/desktop_notifications.ts
@@ -1,0 +1,26 @@
+import {Channel} from '@mattermost/types/channels';
+import {Post} from '@mattermost/types/posts';
+
+import {DesktopNotificationArgs, Store} from 'src/types/mattermost-webapp';
+
+import {ringingEnabled} from 'src/selectors';
+import {isDmGmChannel} from 'src/utils';
+
+export function desktopNotificationHandler(store: Store, post: Post, channel: Channel, args: DesktopNotificationArgs): {error?: string, args?: DesktopNotificationArgs} {
+    if (args.notify) {
+        // Calls will notify if:
+        //  1. it's a custom_calls post (call has started)
+        //  2. in a DM or GM channel
+        //  3. calls is enabled and is v0.17.0+ (it is if this is running)
+        //  4. calls ringing is enabled on the server
+
+        // @ts-ignore our imported webapp types are old
+        if (post.type === 'custom_calls' &&
+            isDmGmChannel(channel) &&
+            ringingEnabled(store.getState())) {
+            return {args: {...args, notify: false}};
+        }
+    }
+
+    return {args};
+}
